MAINT: Consolidate OpenAI response handling
